feat: retire high-priority structural debt in dashboard backend hotspots
